+++ src/parser/ELR/DFA/state/state.ts
@@ -1,4 +1,4 @@
-import type { ILexer } from "../../../../lexer";
+import type { ILexer, ReadonlyILexer } from "../../../../lexer";
 import type { Logger } from "../../../../logger";
 import type { ASTNode } from "../../../ast";
 import type {
@@ -178,7 +178,7 @@ export class State<
    * Return all the possible results after deduplication.
    */
   tryLex(
-    lexer: Readonly<ILexer<LexerError, LexerKinds>>,
+    lexer: ReadonlyILexer<LexerError, LexerKinds>,
     debug: boolean,
     logger: Logger,
   ): {
@@ -240,7 +240,7 @@ export class State<
     entryNTs: ReadonlySet<string>,
     ignoreEntryFollow: boolean,
     followSets: ReadonlyFollowSets<Kinds | LexerKinds>,
-    lexer: Readonly<ILexer<LexerError, LexerKinds>>,
+    lexer: ReadonlyILexer<LexerError, LexerKinds>,
     cascadeQueryPrefix: string | undefined,
     debug: boolean,
     logger: Logger,