+++ src/parser/ELR/model/builder.ts
@@ -1,4 +1,4 @@
-import type { ILexer } from "../../../lexer";
+import type { ILexer, ReadonlyILexer } from "../../../lexer";
 import type { IParser } from "../../model";
 import type { DFA } from "../DFA";
 import type {
@@ -205,7 +205,9 @@ export interface IParserBuilder<
    * So this function is only useful in TypeScript.
    */
   useLexer<AppendLexerKinds extends string, AppendLexerError>(
-    lexer: ILexer<AppendLexerError, AppendLexerKinds>,
+    lexer:
+      | ILexer<AppendLexerError, AppendLexerKinds>
+      | ReadonlyILexer<AppendLexerError, AppendLexerKinds>,
   ): IParserBuilder<
     ASTData,
     ErrorType,