+++ src/parser/ELR/model/context.ts
@@ -1,4 +1,4 @@
-import type { ILexer } from "../../../lexer";
+import type { ReadonlyILexer } from "../../../lexer";
 import type { ASTNode } from "../../ast";
 import type {
   ASTNodeChildrenSelector,
@@ -42,10 +42,10 @@ export class GrammarRuleContext<
    */
   readonly $$: ASTNodeChildrenSelector<ASTData, ErrorType, Kinds | LexerKinds>;
   /**
-   * Current lexer state. You'd better not modify it.
+   * Current lexer state.
    * If you need to modify it, please use `lexer.clone()` or `lexer.dryClone()`.
    */
-  readonly lexer: ILexer<LexerError, LexerKinds>;
+  readonly lexer: ReadonlyILexer<LexerError, LexerKinds>;
   /**
    * Data of the result AST node.
    * You can set this field, and if the grammar rule is accepted,